// ast/ts/types.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * A specifier marking the accessibility of a declaration.
 */
export type Accessibility = "public" | "private";
export type Expression =
  | {
      identifier: string;
    }
  | {
      "string-literal": string;
    }
  | {
      "numeric-literal": string;
    }
  | {
      "binary-operation": {
        lhs: Expression;
        rhs: Expression;
        op: "+" | "-" | "*" | "/";
      };
    }
  | {
      call: {
        function: Expression;
        arguments: Expression[];
      };
    };
/**
 * A line in a function body
 */
export type Statement =
  | {
      declaration: Declaration;
    }
  | {
      expression: Expression;
    }
  | {
      return: Expression;
    };

/**
 * A single Mag source code file.
 */
export interface Program {
  imports?: Import[];
  declarations: {
    accessibility: Accessibility;
    function: Function;
  }[];
}
/**
 * A single import line.
 */
export interface Import {
  path: string;
  symbols: string[];
}
/**
 * A function declaration.
 */
export interface Function {
  name: string;
  arguments: {
    identifier: string;
    type?: Expression;
  }[];
  returnType?: Expression;
  body: Statement[];
}
export interface Declaration {
  identifier: string;
  type?: Expression;
  value?: Expression;
}